+++ CHANGELOG.md
@@ -2,11 +2,12 @@
 
 Erl2ex is currently pre-alpha software. Expect significant backwards-incompatible changes for the time being.
 
-## v0.1.0 (not yet released)
+## v0.0.10 (not yet released)
 
 *   Support for variables in fun M:F/A.
 *   A macro argument that collides with a reserved word or function name wasn't properly renamed. Fixed.
 *   The "inline" compile option didn't map function names properly. Fixed.
+*   The "nowarn_unused_function" compile option is broken under Elixir. Disabled it during transpilation.
 *   Further simplified conversion of public functions with normally un-deffable names.
 *   Documented the code better.